Add files via upload
The excel file SR Configuration Data.xlsx, provides information around which SRs compartments have dual/triple validation.  

 

The aim of the task is to give one value for the SR so the logic would be:

 

Check if the dual/triple validation values are within 5%, if they are take the average to find the level of the compartment. Then take the average across compartments.

 

If not,  for dual work out a way to select the correct one and flag, for triple two might still be correct therefore take the average of those and if one is correct take that. To work out the correct values we discussed we could potentially use the rolling average value.